# server/docs/swagger.yaml
basePath: /api
definitions:
  github_com_xinbreak_movie-web-app_internal_models.Comment:
    properties:
      content:
        type: string
      created_at:
        type: string
      parent:
        $ref: '#/definitions/github_com_xinbreak_movie-web-app_internal_models.Comment'
      parent_id:
        type: string
      replies:
        items:
          $ref: '#/definitions/github_com_xinbreak_movie-web-app_internal_models.Comment'
        type: array
      user:
        $ref: '#/definitions/github_com_xinbreak_movie-web-app_internal_models.User'
      user_id:
        type: string
      video:
        $ref: '#/definitions/github_com_xinbreak_movie-web-app_internal_models.Video'
      video_id:
        type: string
    type: object
  github_com_xinbreak_movie-web-app_internal_models.User:
    properties:
      avatar_url:
        type: string
      comments:
        items:
          $ref: '#/definitions/github_com_xinbreak_movie-web-app_internal_models.Comment'
        type: array
      email:
        type: string
      password:
        type: string
      username:
        type: string
      videos:
        items:
          $ref: '#/definitions/github_com_xinbreak_movie-web-app_internal_models.Video'
        type: array
    type: object
  github_com_xinbreak_movie-web-app_internal_models.Video:
    properties:
      created_at:
        type: string
      description:
        type: string
      duration:
        type: integer
      id:
        type: string
      status:
        type: string
      thumbnail_url:
        type: string
      title:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
      views_count:
        type: integer
    type: object
  github_com_xinbreak_movie-web-app_internal_models_dtos.CreateCommentRequest:
    properties:
      content:
        maxLength: 5000
        minLength: 1
        type: string
      parent_id:
        type: string
      video_id:
        type: string
    required:
    - content
    - video_id
    type: object
...
  github_com_xinbreak_movie-web-app_internal_models_dtos.UpdateCommentRequest:
    properties:
      content:
        maxLength: 5000
        minLength: 1
        type: string
    required:
    - content
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API Server for Movie Application
  title: movie-web-app API
  version: "1.0"
paths:
...
  /comments:
    post:
      consumes:
      - application/json
      description: Create a new comment for a video
      parameters:
      - description: Comment creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_xinbreak_movie-web-app_internal_models_dtos.CreateCommentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_xinbreak_movie-web-app_internal_models.Comment'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Create a new comment
      tags:
      - comments
  /comments/{id}:
    delete:
      description: Delete a comment by ID (only author can delete)
      parameters:
      - description: Comment ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Delete a comment
      tags:
      - comments
    get:
      description: Get a single comment by its ID with replies
      parameters:
      - description: Comment ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_xinbreak_movie-web-app_internal_models.Comment'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get a comment by ID
      tags:
      - comments
    put:
      consumes:
      - application/json
      description: Update an existing comment by ID
      parameters:
      - description: Comment ID
        in: path
        name: id
        required: true
        type: string
      - description: Comment update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_xinbreak_movie-web-app_internal_models_dtos.UpdateCommentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_xinbreak_movie-web-app_internal_models.Comment'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Update a comment
      tags:
      - comments
  /comments/{id}/replies:
    get:
      description: Get paginated replies for a specific comment
      parameters:
      - description: Comment ID
        in: path
        name: id
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get comment replies
      tags:
      - comments
...
  /videos/{video_id}/comments:
    get:
      description: Get paginated comments for a specific video
      parameters:
      - description: Video ID
        in: path
        name: video_id
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get video comments
      tags:
      - comments
swagger: "2.0"